refactoring methods
